[ovium] fix bincode deserialization unexpectedeof error
The error happened when remote commands stderr / stdout were containing
a new line character (\n). To read from the Unix socket, BufRead's
read_until() method is used, with (\n) as the delimitor.
This means it is required to escape new line characters before serializing
data by the server and to unescape it when data is deserialized
by the client.